Update commands to show error message if field does not exist

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -4,13 +4,16 @@ import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_IC_NUMBER;
 
 import java.util.List;
+import java.util.logging.Logger;
 
+import seedu.address.commons.core.LogsCenter;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.patient.IcNumber;
 import seedu.address.model.patient.Patient;
+import seedu.address.model.patient.exceptions.PatientWithFieldNotFoundException;
 
 /**
  * Deletes a patient identified using it's displayed index from the address book.
@@ -27,17 +30,23 @@ public class DeleteCommand extends Command {
 
     public static final String MESSAGE_DELETE_PATIENT_SUCCESS = "Deleted Patient: %1$s";
     private final IcNumber icNumber;
+    private final Logger logger = LogsCenter.getLogger(getClass());
 
     public DeleteCommand(IcNumber icNumber) {
         this.icNumber = icNumber;
     }
 
     @Override
-    public CommandResult execute(Model model) throws CommandException {
+    public CommandResult execute(Model model) throws  PatientWithFieldNotFoundException {
         requireNonNull(model);
         List<Patient> lastShownList = model.getFilteredPatientList();
 
-        Patient patientToDelete = model.getPatient(icNumber, lastShownList);
+        Patient patientToDelete = model.getPatient(icNumber, currentPatientList);
+
+        if (patientToDelete == null) { // no patient with that IC
+            throw new PatientWithFieldNotFoundException("Ic Number : " + icNumber.value);
+        }
+        logger.info("DeleteCommand : " + this + "\nsuccessfully executed");
         model.deletePatient(patientToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PATIENT_SUCCESS, Messages.format(patientToDelete)));
     }
